flatfield/pyproject.toml
[tool.poetry]
name = "flatfield"
version = "0.1.0"
description = ""
authors = ["Jacob Waksmacki <jwaksmac@med.umich.edu>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.22.2"
Pillow = "^9.0.1"
matplotlib = "^3.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
